Add real suggestion to option_map_unwrap_or
